requirements-api.txt
fastapi
uvicorn
pymongo
requests
pydantic
pandas
numpy
scikit-learn
mlflow
dvc
python-dotenv
pytest
jupyter
matplotlib
seaborn
detect-delimiter
passlib
bcrypt
pyjwt
python-multipart
joblib